+++ packages/vscode-extension/CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
 
-All notable changes to the Google Ads Query Language (GAQL) VSCode Extension will be documented in this file.
+## [0.1.2] - 2025-10-25
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## Fixed
+
+- Bug with template literal handling in VS Code extension validation
+- Localization issue causing incorrect display of error messages
 
 ## [0.1.1] - 2025-10-25
 
@@ -42,5 +44,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+[0.1.2]: https://github.com/kage1020/google-ads-query-language/releases/tag/vscode-extension-v0.1.2
 [0.1.1]: https://github.com/kage1020/google-ads-query-language/releases/tag/vscode-extension-v0.1.1
 [0.1.0]: https://github.com/kage1020/google-ads-query-language/releases/tag/vscode-extension-v0.1.0